[update] Use new FileSelector to install runtime(to fix crash on colorOS). Add a warning to the runtime cleanup feature.
