+++ assets/src/main/res/values-en/strings.xml
@@ -281,5 +281,6 @@
 	<string name="tips_redirecting">Redirecting...</string>
 	<string name="tips_please_download_authlib_injector">Missing authlib-injector library. Do you want to download it now?</string>
 	<string name="tips_error">Error: %s</string>
+	<string name="tips_are_you_sure_to_delete_runtime">This operation will clear the runtime you have installed. Are you sure you want to continue?</string>
 
 </resources>
